requirements.txt
Pillow==9.5.0
Pillow==10.0.0
streamlit==1.24.1
